Create bellabeat_analysis.sql
We only received sleep data for 413 customer days. However, we received activity data for 940 customer days. There are a couple ways to interpret this. It is possible that the sleep data is incomplete. It’s also possible that the customers were not wearing their devices to sleep. I can think of a couple of reasons they might not do this. One is that they charge their device overnight. The other is that the device is uncomfortable to wear to sleep. I think if the business could both extended battery life and make the devices more comfortable, specifically for sleeping, they could take advantage of that gap.